+++ packages/data-connect/CHANGELOG.md
@@ -1,3 +1,5 @@
-# Unreleased
-
+# @firebase/data-connect
+## UNRELEASED
+* Updated reporting to use @firebase/data-connect instead of @firebase/connect
+* Added functionality to retry queries and mutations if the server responds with UNAUTHENTICATED.
 * Moved `validateArgs` to core SDK